Move IPC types into herder_api module
