+++ src/herder_daemon/mod.rs
@@ -8,23 +8,21 @@ use tracing::info;
 use tracing_unwrap::ResultExt;
 
 use crate::{
-    herder_daemon::ipc::{TopLevelHerdEvent, WriteVerifyAction},
+    herder_api::{StartHerd, TopLevelHerdEvent, write_verify::WriteVerifyAction},
     ipc_common::{read_msg_async, write_msg},
 };
 
-pub mod ipc;
 mod writer_process;
 
 pub async fn main() {
     loop {
-        let msg =
-            match read_msg_async::<ipc::StartHerd<WriteVerifyAction>>(tokio::io::stdin()).await {
-                Ok(d) => d,
-                Err(e) => {
-                    tracing::info!("Error received on stdin, quitting: {e}");
-                    return;
-                }
-            };
+        let msg = match read_msg_async::<StartHerd<WriteVerifyAction>>(tokio::io::stdin()).await {
+            Ok(d) => d,
+            Err(e) => {
+                tracing::info!("Error received on stdin, quitting: {e}");
+                return;
+            }
+        };
         info!(?msg, "Received StartAction request");
 
         let child = writer_process::spawn_writer(
