+++ src/herder_facade/facade.rs
@@ -1,7 +1,7 @@
 use super::client::LazyHerderClient;
 use super::{HerdHandle, HerderFacade, StartWriterError};
 use crate::escalation::run_escalate;
-use crate::herder_daemon::ipc::{HerdAction, HerdEvent, TopLevelHerdEvent};
+use crate::herder_api::{HerdAction, HerdEvent, TopLevelHerdEvent};
 use crate::herder_facade::DaemonError;
 use crate::herder_facade::client::{HerderClient, HerderClientFactory, RawHerderClient};
 use crate::ipc_common::read_msg_async;
